+++ core/database/src/androidMain/AndroidManifest.xml
