+++ core/database/src/androidMain/kotlin/com/google/samples/apps/nowinandroid/core/database/di/DatabaseModule.android.kt
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.database.di
+
+import android.content.Context
+import androidx.sqlite.db.SupportSQLiteDatabase
+import app.cash.sqldelight.async.coroutines.synchronous
+import app.cash.sqldelight.db.QueryResult
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.db.SqlSchema
+import app.cash.sqldelight.driver.android.AndroidSqliteDriver
+import app.cash.sqldelight.driver.android.AndroidSqliteDriver.Callback
+import org.koin.android.ext.koin.androidContext
+import org.koin.core.module.Module
+import org.koin.dsl.module
+
+internal actual val driverModule: Module
+    get() = module {
+        single<SqlDriver> {
+            val schema: SqlSchema<QueryResult.AsyncValue<Unit>> =
+                get<SqlSchema<QueryResult.AsyncValue<Unit>>>()
+            val context: Context = androidContext()
+            val synchronousSchema = schema.synchronous()
+            AndroidSqliteDriver(
+                schema = synchronousSchema,
+                context = context,
+                name = "nia-database.db",
+                callback = object : Callback(synchronousSchema) {
+                    override fun onOpen(db: SupportSQLiteDatabase) {
+                        db.setForeignKeyConstraintsEnabled(true)
+                    }
+                },
+            )
+        }
+    }
