+++ core/database/src/androidUnitTest/kotlin/com/google/samples/apps/nowinandroid/core/database/BaseDatabaseTest.android.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.database
+
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+import java.util.Properties
+
+/**
+ * Init driver for each platform. Should *always* be called to setup test
+ */
+actual suspend fun createDriver(): SqlDriver {
+    return JdbcSqliteDriver(
+        url = JdbcSqliteDriver.IN_MEMORY,
+        properties = Properties().apply { put("foreign_keys", "true") },
+    ).also { NiaDatabase.Schema.create(it).await() }
+}
