+++ core/database/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/database/mapper/RecentSearchQueryMapper.kt
