+++ core/database/src/commonMain/kotlin/com/google/samples/apps/nowinandroid/core/database/model/RecentSearchQueryEntity.kt
