+++ core/database/src/commonMain/sqldelight/com/google/samples/apps/nowinandroid/core/database/RecentSearchQuery.sq
