+++ core/database/src/commonTest/kotlin/com/google/samples/apps/nowinandroid/core/database/BaseDatabaseTest.kt
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.database
+
+import app.cash.sqldelight.db.SqlDriver
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.test.runTest
+
+/**
+ * Init driver for each platform. Should *always* be called to setup test
+ */
+expect suspend fun createDriver(): SqlDriver
+
+/**
+ * A test function that creates a [NiaDatabase] instance and runs the provided block of code
+ * within a coroutine scope.
+ *
+ * @param block The block of code to be executed with the created [NiaDatabase] instance.
+ */
+fun databaseTest(block: suspend CoroutineScope.(NiaDatabase) -> Unit) = runTest {
+    val driver = createDriver()
+    block(NiaDatabase(driver))
+}
