+++ core/database/src/commonTest/kotlin/com/google/samples/apps/nowinandroid/core/database/dao/NewsResourceDaoTest.kt
@@ -1,0 +1,300 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.database.dao
+
+import com.google.samples.apps.nowinandroid.core.database.databaseTest
+import com.google.samples.apps.nowinandroid.core.database.model.NewsResourceEntity
+import com.google.samples.apps.nowinandroid.core.database.model.NewsResourceTopicCrossRef
+import com.google.samples.apps.nowinandroid.core.database.model.TopicEntity
+import com.google.samples.apps.nowinandroid.core.database.model.asExternalModel
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.test.runTest
+import kotlinx.datetime.Instant
+import kotlin.test.BeforeTest
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+internal class NewsResourceDaoTest {
+
+    private lateinit var newsResourceDao: NewsResourceDao
+    private lateinit var topicDao: TopicDao
+
+    @BeforeTest
+    fun setup() = databaseTest {
+        newsResourceDao = NewsResourceDaoImpl(it, Dispatchers.Unconfined)
+        topicDao = TopicDaoImpl(it, Dispatchers.Unconfined)
+    }
+
+    @Test
+    fun getNewsResources_allEntries_areOrderedByPublishDateDesc() = runTest {
+        val newsResourceEntities = listOf(
+            testNewsResource(
+                id = "0",
+                millisSinceEpoch = 0,
+            ),
+            testNewsResource(
+                id = "1",
+                millisSinceEpoch = 3,
+            ),
+            testNewsResource(
+                id = "2",
+                millisSinceEpoch = 1,
+            ),
+            testNewsResource(
+                id = "3",
+                millisSinceEpoch = 2,
+            ),
+        )
+        newsResourceDao.upsertNewsResources(
+            newsResourceEntities,
+        )
+
+        val savedNewsResourceEntities = newsResourceDao.getNewsResources()
+            .first()
+
+        assertEquals(
+            listOf(3L, 2L, 1L, 0L),
+            savedNewsResourceEntities.map {
+                it.asExternalModel().publishDate.toEpochMilliseconds()
+            },
+        )
+    }
+
+    @Test
+    fun getNewsResources_filteredById_areOrderedByDescendingPublishDate() = runTest {
+        val newsResourceEntities = listOf(
+            testNewsResource(
+                id = "0",
+                millisSinceEpoch = 0,
+            ),
+            testNewsResource(
+                id = "1",
+                millisSinceEpoch = 3,
+            ),
+            testNewsResource(
+                id = "2",
+                millisSinceEpoch = 1,
+            ),
+            testNewsResource(
+                id = "3",
+                millisSinceEpoch = 2,
+            ),
+        )
+        newsResourceDao.upsertNewsResources(
+            newsResourceEntities,
+        )
+
+        val savedNewsResourceEntities = newsResourceDao.getNewsResources(
+            useFilterNewsIds = true,
+            filterNewsIds = setOf("3", "0"),
+        )
+            .first()
+
+        assertEquals(
+            listOf("3", "0"),
+            savedNewsResourceEntities.map {
+                it.entity.id
+            },
+        )
+    }
+
+    @Test
+    fun getNewsResources_filteredByTopicId_areOrderedByDescendingPublishDate() = runTest {
+        val topicEntities = listOf(
+            testTopicEntity(
+                id = "1",
+                name = "1",
+            ),
+            testTopicEntity(
+                id = "2",
+                name = "2",
+            ),
+        )
+        val newsResourceEntities = listOf(
+            testNewsResource(
+                id = "0",
+                millisSinceEpoch = 0,
+            ),
+            testNewsResource(
+                id = "1",
+                millisSinceEpoch = 3,
+            ),
+            testNewsResource(
+                id = "2",
+                millisSinceEpoch = 1,
+            ),
+            testNewsResource(
+                id = "3",
+                millisSinceEpoch = 2,
+            ),
+        )
+        val newsResourceTopicCrossRefEntities = topicEntities.mapIndexed { index, topicEntity ->
+            NewsResourceTopicCrossRef(
+                newsResourceId = index.toString(),
+                topicId = topicEntity.id,
+            )
+        }
+
+        topicDao.insertOrIgnoreTopics(
+            topicEntities = topicEntities,
+        )
+        newsResourceDao.upsertNewsResources(
+            newsResourceEntities,
+        )
+        newsResourceDao.insertOrIgnoreTopicCrossRefEntities(
+            newsResourceTopicCrossRefEntities,
+        )
+
+        val filteredNewsResources = newsResourceDao.getNewsResources(
+            useFilterTopicIds = true,
+            filterTopicIds = topicEntities
+                .map(TopicEntity::id)
+                .toSet(),
+        ).first()
+
+        assertEquals(
+            listOf("1", "0"),
+            filteredNewsResources.map { it.entity.id },
+        )
+    }
+
+    @Test
+    fun getNewsResources_filteredByIdAndTopicId_areOrderedByDescendingPublishDate() = runTest {
+        val topicEntities = listOf(
+            testTopicEntity(
+                id = "1",
+                name = "1",
+            ),
+            testTopicEntity(
+                id = "2",
+                name = "2",
+            ),
+        )
+        val newsResourceEntities = listOf(
+            testNewsResource(
+                id = "0",
+                millisSinceEpoch = 0,
+            ),
+            testNewsResource(
+                id = "1",
+                millisSinceEpoch = 3,
+            ),
+            testNewsResource(
+                id = "2",
+                millisSinceEpoch = 1,
+            ),
+            testNewsResource(
+                id = "3",
+                millisSinceEpoch = 2,
+            ),
+        )
+        val newsResourceTopicCrossRefEntities = topicEntities.mapIndexed { index, topicEntity ->
+            NewsResourceTopicCrossRef(
+                newsResourceId = index.toString(),
+                topicId = topicEntity.id,
+            )
+        }
+
+        topicDao.insertOrIgnoreTopics(
+            topicEntities = topicEntities,
+        )
+        newsResourceDao.upsertNewsResources(
+            newsResourceEntities,
+        )
+        newsResourceDao.insertOrIgnoreTopicCrossRefEntities(
+            newsResourceTopicCrossRefEntities,
+        )
+
+        val filteredNewsResources = newsResourceDao.getNewsResources(
+            useFilterTopicIds = true,
+            filterTopicIds = topicEntities
+                .map(TopicEntity::id)
+                .toSet(),
+            useFilterNewsIds = true,
+            filterNewsIds = setOf("1"),
+        ).first()
+
+        assertEquals(
+            listOf("1"),
+            filteredNewsResources.map { it.entity.id },
+        )
+    }
+
+    @Test
+    fun deleteNewsResources_byId() =
+        runTest {
+            val newsResourceEntities = listOf(
+                testNewsResource(
+                    id = "0",
+                    millisSinceEpoch = 0,
+                ),
+                testNewsResource(
+                    id = "1",
+                    millisSinceEpoch = 3,
+                ),
+                testNewsResource(
+                    id = "2",
+                    millisSinceEpoch = 1,
+                ),
+                testNewsResource(
+                    id = "3",
+                    millisSinceEpoch = 2,
+                ),
+            )
+            newsResourceDao.upsertNewsResources(newsResourceEntities)
+
+            val (toDelete, toKeep) = newsResourceEntities.partition { it.id.toInt() % 2 == 0 }
+
+            newsResourceDao.deleteNewsResources(
+                toDelete.map(NewsResourceEntity::id),
+            )
+
+            assertEquals(
+                toKeep.map(NewsResourceEntity::id)
+                    .toSet(),
+                newsResourceDao.getNewsResources().first()
+                    .map { it.entity.id }
+                    .toSet(),
+            )
+        }
+}
+
+private fun testTopicEntity(
+    id: String = "0",
+    name: String,
+) = TopicEntity(
+    id = id,
+    name = name,
+    shortDescription = "",
+    longDescription = "",
+    url = "",
+    imageUrl = "",
+)
+
+private fun testNewsResource(
+    id: String = "0",
+    millisSinceEpoch: Long = 0,
+) = NewsResourceEntity(
+    id = id,
+    title = "",
+    content = "",
+    url = "",
+    headerImageUrl = "",
+    publishDate = Instant.fromEpochMilliseconds(millisSinceEpoch),
+    type = "Article 📚",
+)
