+++ core/database/src/nativeTest/kotlin/com/google/samples/apps/nowinandroid/core/database/BaseDatabaseTest.native.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.samples.apps.nowinandroid.core.database
+
+import app.cash.sqldelight.async.coroutines.synchronous
+import app.cash.sqldelight.db.SqlDriver
+import app.cash.sqldelight.driver.native.NativeSqliteDriver
+import co.touchlab.sqliter.DatabaseConfiguration
+import com.google.samples.apps.nowinandroid.core.database.NiaDatabase.Companion.Schema
+
+/**
+ * Init driver for each platform. Should *always* be called to setup test
+ */
+actual suspend fun createDriver(): SqlDriver {
+    return NativeSqliteDriver(
+        schema = Schema.synchronous(),
+        name = "nia-database-test.db",
+        onConfiguration = { config: DatabaseConfiguration ->
+            config.copy(
+                extendedConfig = DatabaseConfiguration.Extended(foreignKeyConstraints = true),
+            )
+        },
+    )
+}
